+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reindex membership.
 - Allow `multiple` option for algorithms that allow when using MATLAB graphs.
 - Options for storing results as node and edge attributes and using stored attributes in other functions.
+- Support for using MATLAB RNG in igraph and set it as default.
 
 ### Changed